add support for curl
also gather the generic parsing logic like the browser and crawler logic
